refactor

+++ uss.h
@@ -5,6 +5,6 @@ void uss_init(uint8_t trigger_pin, uint8_t echo_pin);
 
 // send and get duration of a pulse
 // going out of range of timeout will return 0
-uint64_t get_pulse_duration(uint8_t trigger_pin, uint8_t echo_pin, uint64_t timeout);
+uint get_pulse_duration(uint8_t trigger_pin, uint8_t echo_pin, uint64_t timeout);
 
 #endif
